requirements.txt
pillow>=10.0.0
mss>=9.0.0
pyautogui>=0.9.54
anthropic>=0.18.0
imagehash>=4.3.1
keyboard>=0.13.5
pynput>=1.7.6
